% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_mar_mat.R
\name{comment_matrix}
\alias{comment_matrix}
\title{Compute feature matrix for toxic/
for toxic and non-toxic comments classification problem}
\usage{
comment_matrix(commentData, key.word.list, id = "id",
  colname = "comment_text", progress = TRUE)
}
\arguments{
\item{commentData}{A dataframe of input comment data}

\item{key.word.list}{A vector of keyword list for which the the number of total count
in each comment will be computed}

\item{colname}{Column's name for comment text. By default is 'comment_text'}

\item{progress}{Display the progress bar if TRUE}
}
\value{
The function will return a sparse matrix whose column is the word count of
the corresponding keyword in each comment
}
\description{
This function will compute the feature (word count) matrix for a comment dataset given an input
keyword list
}
\examples{
out <- Marginal_Freq(train.data, train.data$toxic)
comment_matrix(train.data, out$keyword)

}
