% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Data preprocessing}
\usage{
preprocess(traindata, testdata, prep_fun = tolower, tok_fun = tok_fun)
}
\arguments{
\item{traindata}{a matrix, list or data frame for training, with one of its columns named "comment_text".}

\item{testdata}{a matrix, list or data frame for test, with one of its columns named "comment_text".}

\item{prep_fun}{a function which takes chunk of character vectors and does all pre-processing.}

\item{tok_fun}{a function which takes a character vector from preprocessor, split it into tokens and returns a list of character vectors.}
}
\value{
a list with the following elements:
\itemize{
\item{the document-term matrix for training data}
\item{the document-term matrix for test data}
}
}
\description{
This function converts raw comments data to a document-term matrix that can be used for classification.
By using up to 3-grams and then pruning vocabulary, it can significantly reduce the dimension.
}
\examples{
\dontrun{
data("train.data")
data("test.data")
N.train = 1000
N.test = 100
data.pre <- preprocess(traindata = train.data[1:N.train, ], testdata = test.data[1:N.test, ])
train.dtm <- data.pre$train.dtm
test.dtm  <- data.pre$test.dtm
}

}
