% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.data}
\alias{test.data}
\title{Test data}
\format{A data frame with 153164 rows and 2 variables:
\describe{
\item{id}{Comment ID}
\item{comment_text}{Raw Wikipedia comments}
}}
\source{
\url{https://www.kaggle.com/c/jigsaw-toxic-comment-classification-challenge/data}
}
\usage{
test.data
}
\description{
This is the test set containing comments from approximately 153k subjects.
}
\keyword{datasets}
