% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_labels}
\alias{test_labels}
\title{Test labels}
\format{A data frame with 153164 rows and 7 variables:
\describe{
\item{id}{Comment ID}
\item{toxic}{label for type "toxic"}
\item{severe_toxic}{label for type "severe_toxic"}
\item{obscene}{label for type "obscene"}
\item{threat}{label for type "threat"}
\item{insult}{label for type "insult"}
\item{identity_hate}{label for type "identity_hate"}
}}
\source{
\url{https://www.kaggle.com/c/jigsaw-toxic-comment-classification-challenge/data}
}
\usage{
test_labels
}
\description{
This data set contains labels for the test data. Value of -1 indicates it was not used for scoring.
}
\keyword{datasets}
