% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{tok_fun}
\alias{tok_fun}
\title{Tokenization function for string splitting}
\usage{
tok_fun(x)
}
\arguments{
\item{x}{character vector}
}
\value{
a list of character vectors, each element of the list contains vector of tokens.
}
\description{
This function improves the simple function \strong{word_tokenizer} in \strong{\emph{text2vec}} package by applying stemming after tokenization.
}
\examples{
doc = c("I feel happy", "I get a feeling of happiness", "She feels happy")
tok_fun(doc)

}
