% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train.data}
\alias{train.data}
\title{Training data}
\format{A data frame with 159571 rows and 8 variables:
\describe{
\item{id}{Comment ID}
\item{comment_text}{Raw Wikipedia comments}
\item{toxic}{binary label for type "toxic"}
\item{severe_toxic}{binary label for type "severe_toxic"}
\item{obscene}{binary label for type "obscene"}
\item{threat}{binary label for type "threat"}
\item{insult}{binary label for type "insult"}
\item{identity_hate}{binary label for type "identity_hate"}
}}
\source{
\url{https://www.kaggle.com/c/jigsaw-toxic-comment-classification-challenge/data}
}
\usage{
train.data
}
\description{
This is the training set containing comments with their binary labels.
}
\keyword{datasets}
