% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Data preprocessing}
\usage{
preprocess(traindata, testdata, trainid = "id", testid = "id",
  traintext = "comment_text", testtext = "comment_text",
  prep_fun = tolower, tok_fun = UmbrellaAcademy::tok_fun,
  term_count.min = 10)
}
\arguments{
\item{traindata}{a data frame or data table for training, two columns containing comment ID and comment text must be included.}

\item{testdata}{a data frame or data table for test, two columns containing comment ID and comment text must be included.}

\item{trainid}{column's name for comment ID of training data (default is "id"), quotation marks needed.}

\item{testid}{column's name for comment ID of test data (default is "id"), quotation marks needed.}

\item{traintext}{column's name for comment text of training data (default is "comment_text"), quotation marks needed.}

\item{testtext}{column's name for comment text of test data (default is "comment_text"), quotation marks needed.}

\item{prep_fun}{a function which takes chunk of character vectors and does all pre-processing (default is \strong{tolower}).}

\item{tok_fun}{a function which takes a character vector from preprocessor, split it into tokens and returns a list of character vectors (default is \strong{tok_fun} in \strong{\emph{UmbrellaAcademy}} package).}

\item{term_count.min}{minimum number of occurences for a term over all comments (default is 10).}
}
\value{
a list with the following elements:
\describe{
\item{train.dtm}{the document-term matrix for training data}
\item{raw.test.dtm}{the raw document-term matrix for test data}
\item{it_test}{itoken iterator for test data}
\item{raw.vocab}{vocabulary object for training data}
}
}
\description{
This function converts raw comments data to a document-term matrix that can be used for classification.
By using up to 3-grams and then pruning vocabulary, it can significantly reduce the dimension.
}
\examples{
\dontrun{
data("train.data")
data("test.data")
N.train = 1000
N.test = 100
data.pre <- preprocess(traindata = train.data[1:N.train, ], testdata = test.data[1:N.test, ])
train.dtm <- data.pre$train.dtm
test.dtm  <- data.pre$test.dtm
}

}
